refactor: enhance routing and documentation structure
- Implemented additional redirects in next.config.ts for better navigation.
- Updated index.mdx links to match the new routing framework.
- Modified metadata generation in page.tsx to correspond with revised URL paths.
- Adjusted base URL in source.ts for improved application consistency.
